Updated RtpDump Reader and Writer to ensure they are writing compatible files that of rtpdump.
Updated tests to show how to read Header and Payload format files if required at runtime (Not required for people who do not use rtptools, this is just for completeness)

Updated Unit Test to show that Reading of rtpdump files created by the standard rtpdump toolset work and that the library can write files which are in the same format.

Binary (Header, Payload), ASCII and Hex are complete!!!

Short form needs to be implemented to have 100% of the feautres of rtpdump but that will be later if required.

Added Utility Function List<T> AddRange with start and length.

Added Utility Function GetBytes(string)

git-svn-id: https://net7mma.svn.codeplex.com/svn@99515 4ef1bfeb-0a34-4d02-9305-240e5b07a168