A bug related to the RtpMarker in the CommonHeaderBits class has been fixed. (Thanks to Aaron @ Codeplex)
Added 'R' To ASCII.

When creating a SourceList instance the `buffer` length property was not checked and would cause problems when attempting completing a packets.

The RtcpHeader class was not using the PointerToLast6Bytes property propertly and as a result packets were being mangled when received and interpreted incorreclty.

RtpClient

Certain properties were not being checked for disposition and as a result the instance would throw an exception when receiving data in the 'SendReceive' function.

The wonderful block of exception handling ode in the 'SendReceive' function has been reduced.

I removed the `OutOfBandInline` and  opted to use `KeepAliveValues` since the latter doesn't work with the former, Nagle is still disabled.

Packets events were being raised twice before and and since that really isn't required the semantic has been changed to only raise them once.

I was playing around with a pattern which only exposes the header of the packet recieved but this experiment did not make it into this release.

Rtp and Rtcp packet handling code has been changed to be as minimal as possible.

Handlers are now checked against null because there are no default handlers, it may make sense to use a lambda / sentinel to prevent the null check however there are bigger fish to fry.

The RtpClient no longer uses to / from compound bytes methods when sending compound packets.

The test program now shows the time playing and time remaning in the `media`...



git-svn-id: https://net7mma.svn.codeplex.com/svn@106575 4ef1bfeb-0a34-4d02-9305-240e5b07a168