SourceDescription - now use UTF8 rather than System.Default encoding.
RtpClient -

Set the default Send and Receive timeout on the RtpClient.

Disabled forcing the use of OverlappedIO by default.

Removed the check for specific RTCP payload types in the HandleIncomingRtcpPacket function.

Changed when reports for Rtcp were being sent, might choose to integrate this into the SendKeepAlive process at some point.

Jitter is always updated when a RtpPacket is sent.

Fixed the RtcpBandwidthExceeded calculation


RtspServer -

Updates for 2.0 Draft. (my notes on the current status are not included at this time)

Change in methodology for allowing TCP, the RFC4571 'TCP/RTP/AVP' token causes VLC to choke and die.

Verified that a UDP stream can be re-sourced over TCP and watched with VLC.

Verified that UDP streaming is working.

Changed test program to facilite these tests.

The test program also now saves the SDP of the currently playing media while the test is running and deletes it afterwards.











git-svn-id: https://net7mma.svn.codeplex.com/svn@106615 4ef1bfeb-0a34-4d02-9305-240e5b07a168