Targeted .Net 4.5.2 Runtime by default (Testing RyuJit, will help with MPEG Encoding)
Renamed RFC3016Stream to RFC2250Stream, RFC3016Stream will follow.

Added ISource, ISink, IMediaStream Interfaces and the RtpSink class.

Added the ability to optionally skip profile headers when RtpFrame's are assembled.

Added RtpPacket consturctor to makes headers on the fly.

Started basis for other types of Transports e.g. Rtmp or MPEG TS, ClientSession is the only remaining class which needs to be revised and then some small changes in the RtspServer and RtspSource classes.

Moved RFC2435Frame to RFC2435Stream class, all future Packetization and Depacketization implemenations will be encapsulated in the supporting Stream class, this will allow seperation of types which are not required in certain implementations.

Fixed RtpCounter calculations, was taking all bytes now only counts the payload bytes when sending RtcpReports per the RFC and has a TotalBytes property which calculates the overhead of the RtpHeaders for each packet sent on the fly when needed.



git-svn-id: https://net7mma.svn.codeplex.com/svn@108490 4ef1bfeb-0a34-4d02-9305-240e5b07a168