Ensured Server Works with VLC, Quicktime etc.
Moved RFC3550 back to Media namespace.

ClientSession now returns the correct port in the SDP after the session has began playing (for subsequent describes)

ClientSession - Events are only attached once so Pause works and so does Disabling and Enabling a track in VLC.

MediaFileStream - Added GetCompatbileImplementation to help find MediaFileStream for a unknown type of file.

git-svn-id: https://net7mma.svn.codeplex.com/svn@109999 4ef1bfeb-0a34-4d02-9305-240e5b07a168