*Stable Nightly Container Release*
Utility - ContainsBytes now respects given count.

RtspClient - Does not wait for response for play or teardown.

Always uses the underlying transport to parse requests when possible.

Rtsp Unreliable Support Working again.

*Thanks to exterminator @ codeplex for finding this and pointing it out*

RtspMessage - Fixed a bug when partial messages were received. The entire recieved data is used as if it was the entire message.

If the message does not contain at least a first line it is considered invalid.

If there is a content-length but more headers are present in the buffer the additional headers will be seen as part of the body.

git-svn-id: https://net7mma.svn.codeplex.com/svn@110649 4ef1bfeb-0a34-4d02-9305-240e5b07a168