*Stable Container Release*
[Bug Fixes, Compliance, Performance]

- This is the last release scheduled for 'Container' -

Revised some of the Sdp classes.

MediaDescription now has a line which represents itself.

TimeDescription now has a line which respresents itself.

SessionDescriptionLine is IEnumerable<string>

SessionDescriptionLine gets a new constructor

SessionConnectionLine caches the split for the connection parts.

Properties are named with 'Connection' where possible in the SessionConnectionLine.

RtpClient - When sending a Goodbye the ssrc is inserted into the SourceList even if the id will be the same as the SendersSynchronizationSourceIdentifier. (previously a BlockCount of 1 would be sent but no id's because the SSRC was alreayd indicated by the SendersSynchronizationSourceIdentifier)

Added logging points.

Handled Rtcp Bandwidth parsing better when it was not present.

Revised RTCP Handling

SourceList - Fixed Min Max issue.

RtspClient - Ensures responses were received according to what was sent.

Added disconnect parameter for SendTeardown.

ClientSession - used HasAnyActivity to reduce code.

Program - Revised unit testing procedures.

git-svn-id: https://net7mma.svn.codeplex.com/svn@110925 4ef1bfeb-0a34-4d02-9305-240e5b07a168